Within addEntriesToZip(), which is used for merging multiple CBZ files into one, replaced usage of ZipInputStream with ZipFile.
On Android 15 the usage of ZipInputStream directly as used previously, only retrieves the first file of each CBZ, making this change works both on Android 15 and the previous versions.

To utilize ZipFile, I decided to reuse the copyCbzToCacheAndCloseInputStream() to make a copy of file and decided to refactor it to catch error within the method.