Updated itextpdf library from version 7 to version 9. Updated DEFAULT_MAX_NUMBER_OF_PAGES to 10,000. Cache directory will now be cleared, when starting the process of converting a CBZ to PDF. Increased padding size of files when merging them into one giant CBZ, to support larger single files being merged.
Added new logic for when merging files occurs. If total number of pages are 300 or more, we will split them into multiple pdfs of 300 pages. Then use each of those 300 page pdf to directly copy them into a single pdf, this will limit memory usage to the size of a 300 page pdf at most. When merging is less than 300, previous logic of directly merging in memory will be used.